Add a method to ensure non-repeating digits in code
Repeating digits are currently not fully supported by the computer
breaker AI algorithm that I implemented. A subset of the possible codes
with repeating digits cause unexpected behavior that will be revisited
later. To be clear, several codes with repeating digits can be solved by
the AI. However, I've decided to restrict the code to unique numbers for
now.